Removes pre-RC1 endpoint: `/save_card_to_customer`
This was ~replaced with `/attach_payment_method_to_customer` in #8, and now that all of the
client SDKs have been updated & released to use PaymentMethods, this Sources-only endpoint
is unused.

Since we're not (yet?) tagging releases of the example backend, this means anyone who
deploys this project & tries to use it from an older beta of the clients will be broken,
but there shouldn't be any/many people doing that.